feat(settings): move appearance preferences into user profile
