+++ frontend/src/app/features/settings/user-profile-dialog/user-profile-dialog.component.scss
@@ -66,6 +66,62 @@
   }
 }
 
+.preferences-section {
+  gap: 1rem;
+}
+
+.language-select {
+  width: 11rem;
+  max-width: 100%;
+  min-width: 0;
+}
+
+.preference-group {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.theme-row {
+  display: flex;
+  flex-direction: column;
+  gap: 0.625rem;
+  padding: 0.75rem;
+  background: var(--ground-background);
+  border-radius: 8px;
+  box-shadow: inset 0 0 0 1px var(--border-color);
+}
+
+.theme-row-label {
+  color: var(--text-color-secondary);
+  font-size: var(--text-sm);
+  font-weight: 500;
+  line-height: 1.25rem;
+}
+
+.color-swatches {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.375rem;
+  min-width: 0;
+
+  button {
+    width: 1.5rem;
+    height: 1.5rem;
+    padding: 0;
+    border: 0;
+    border-radius: 50%;
+    cursor: pointer;
+    outline: transparent solid 2px;
+    outline-offset: 1px;
+
+    &:focus-visible,
+    &.active-color {
+      outline-color: var(--primary-color);
+    }
+  }
+}
+
 .form-field {
   display: flex;
   flex-direction: column;
@@ -153,5 +209,11 @@
   p {
     font-size: 1rem;
     color: var(--text-secondary-color);
+  }
+}
+
+@media (width <= 480px) {
+  .language-select {
+    width: 100%;
   }
 }
