+++ frontend/src/app/features/settings/user-profile-dialog/user-profile-dialog.component.ts
@@ -1,4 +1,5 @@
-import {Component, effect, inject} from '@angular/core';
+import {NgClass} from '@angular/common';
+import {Component, DestroyRef, computed, effect, inject} from '@angular/core';
 import {Button} from 'primeng/button';
 import {AbstractControl, FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, ValidationErrors, ValidatorFn, Validators} from '@angular/forms';
 import {InputText} from 'primeng/inputtext';
@@ -7,6 +8,19 @@ import {User, UserService, UserUpdateRequest} from '../user-management/user.serv
 import {MessageService} from 'primeng/api';
 import {DynamicDialogRef} from 'primeng/dynamicdialog';
 import {TranslocoDirective, TranslocoPipe, TranslocoService} from '@jsverse/transloco';
+import {Select} from 'primeng/select';
+import {takeUntilDestroyed, toSignal} from '@angular/core/rxjs-interop';
+import {AVAILABLE_LANGS, LANG_LABELS} from '../../../core/config/transloco-loader';
+import {LANG_STORAGE_KEY} from '../../../core/config/language-initializer';
+import {AppConfigService} from '../../../shared/service/app-config.service';
+import Aura from '../../../shared/layout/theme-palette-extend';
+
+type ColorPalette = Record<string, string>;
+
+interface Palette {
+  name: string;
+  palette: ColorPalette;
+}
 
 export const passwordMatchValidator: ValidatorFn = (control: AbstractControl): ValidationErrors | null => {
   const newPassword = control.get('newPassword');
@@ -27,8 +41,10 @@ export const passwordMatchValidator: ValidatorFn = (control: AbstractControl): V
     ReactiveFormsModule,
     InputText,
     Password,
+    Select,
+    NgClass,
     TranslocoDirective,
-    TranslocoPipe
+    TranslocoPipe,
   ],
   templateUrl: './user-profile-dialog.component.html',
   styleUrls: ['./user-profile-dialog.component.scss']
@@ -39,12 +55,23 @@ export class UserProfileDialogComponent {
   currentUser: User | null = null;
   editUserData: Partial<User> = {};
   changePasswordForm: FormGroup;
+  readonly languageOptions = AVAILABLE_LANGS.map(value => ({
+    value,
+    label: LANG_LABELS[value] ?? value,
+  }));
 
   protected readonly userService = inject(UserService);
+  protected readonly configService = inject(AppConfigService);
   private readonly messageService = inject(MessageService);
   private readonly fb = inject(FormBuilder);
   private readonly dialogRef = inject(DynamicDialogRef);
   private readonly t = inject(TranslocoService);
+  private readonly destroyRef = inject(DestroyRef);
+  protected readonly activeLang = toSignal(this.t.langChanges$, {initialValue: this.t.getActiveLang()});
+  protected readonly selectedPrimaryColor = computed(() => this.configService.appState().primary);
+  protected readonly selectedSurfaceColor = computed(() => this.configService.appState().surface);
+  protected readonly surfaces = this.configService.surfaces;
+  protected readonly primaryColors = this.getPrimaryColors();
 
   constructor() {
     this.changePasswordForm = this.fb.group(
@@ -80,6 +107,43 @@ export class UserProfileDialogComponent {
         email: this.currentUser.email,
       };
     }
+  }
+
+  onLanguageChange(lang: string): void {
+    if (lang === this.activeLang()) return;
+    this.t.load(lang).pipe(takeUntilDestroyed(this.destroyRef)).subscribe({
+      next: () => {
+        this.t.setActiveLang(lang);
+        localStorage.setItem(LANG_STORAGE_KEY, lang);
+      },
+    });
+  }
+
+  updateThemeColor(event: Event, type: 'primary' | 'surface', color: { name: string; palette?: ColorPalette }): void {
+    this.configService.appState.update((state) => ({
+      ...state,
+      [type]: color.name,
+    }));
+    event.stopPropagation();
+  }
+
+  private getPrimaryColors(): Palette[] {
+    const presetPalette = (Aura.primitive ?? {}) as Record<string, ColorPalette>;
+    const colors = [
+      'orange', 'amber', 'yellow', 'lime', 'green', 'emerald',
+      'teal', 'cyan', 'sky', 'blue', 'indigo', 'violet',
+      'purple', 'fuchsia', 'pink', 'rose', 'red',
+      'coralSunset', 'roseBlush', 'melonBlush', 'cottonCandy',
+      'apricotSunrise', 'antiqueBronze', 'butteryYellow', 'vanillaCream',
+      'citrusMint', 'freshMint', 'sagePearl', 'skyBlue', 'periwinkleCream',
+      'pastelRoyalBlue', 'lavenderDream', 'dustyNeutral',
+    ];
+    return [{name: 'noir', palette: {}}].concat(
+      colors.map(name => ({
+        name,
+        palette: presetPalette[name] ?? {},
+      })),
+    );
   }
 
   updateProfile(): void {
