--- frontend/src/app/features/settings/view-preferences-parent/language-preferences/language-preferences.component.scss
@@ -1,41 +1,0 @@
-@use '../../../../shared/styles/settings-shared' as settings;
-
-.settings-card {
-  @include settings.settings-card;
-}
-
-.section-header {
-  @include settings.settings-section-header;
-}
-
-.section-title {
-  @include settings.settings-section-title;
-}
-
-.section-description {
-  @include settings.settings-section-description;
-}
-
-.section-body {
-  @include settings.settings-section-body;
-}
-
-.setting-item {
-  @include settings.settings-item;
-}
-
-.setting-header-standalone {
-  @include settings.settings-item-header-standalone;
-}
-
-.setting-label {
-  @include settings.settings-label;
-}
-
-.setting-options {
-  @include settings.settings-options;
-}
-
-.select-group {
-  @include settings.settings-select-group;
-}
