--- frontend/src/app/features/settings/view-preferences-parent/language-preferences/language-preferences.component.ts
@@ -1,35 +1,0 @@
-import {Component, DestroyRef, inject} from '@angular/core';
-import {FormsModule} from '@angular/forms';
-import {takeUntilDestroyed, toSignal} from '@angular/core/rxjs-interop';
-import {Select} from 'primeng/select';
-import {TranslocoDirective, TranslocoService} from '@jsverse/transloco';
-import {AVAILABLE_LANGS, LANG_LABELS} from '../../../../core/config/transloco-loader';
-import {LANG_STORAGE_KEY} from '../../../../core/config/language-initializer';
-
-@Component({
-  selector: 'app-language-preferences',
-  standalone: true,
-  imports: [FormsModule, Select, TranslocoDirective],
-  templateUrl: './language-preferences.component.html',
-  styleUrl: './language-preferences.component.scss',
-})
-export class LanguagePreferencesComponent {
-  readonly languageOptions = AVAILABLE_LANGS.map(value => ({
-    value,
-    label: LANG_LABELS[value] ?? value,
-  }));
-
-  private readonly t = inject(TranslocoService);
-  private readonly destroyRef = inject(DestroyRef);
-  readonly activeLang = toSignal(this.t.langChanges$, {initialValue: this.t.getActiveLang()});
-
-  onLanguageChange(lang: string): void {
-    if (lang === this.activeLang()) return;
-    this.t.load(lang).pipe(takeUntilDestroyed(this.destroyRef)).subscribe({
-      next: () => {
-        this.t.setActiveLang(lang);
-        localStorage.setItem(LANG_STORAGE_KEY, lang);
-      },
-    });
-  }
-}
