--- frontend/src/app/features/settings/view-preferences-parent/theme-preferences/theme-preferences.component.scss
@@ -1,21 +1,0 @@
-@use '../../../../shared/styles/settings-shared' as settings;
-
-.settings-card {
-  @include settings.settings-card;
-}
-
-.section-header {
-  @include settings.settings-section-header;
-}
-
-.section-title {
-  @include settings.settings-section-title;
-}
-
-.section-description {
-  @include settings.settings-section-description;
-}
-
-.section-body {
-  @include settings.settings-section-body;
-}
