+++ frontend/src/app/features/settings/view-preferences-parent/view-preferences-parent.component.html
@@ -10,11 +10,6 @@
       </p>
     </div>
 
-    <app-language-preferences></app-language-preferences>
-    <div class="section-divider"><hr></div>
-    <app-theme-preferences></app-theme-preferences>
-    <div class="section-divider"><hr></div>
-
     <app-view-preferences></app-view-preferences>
     <div class="section-divider"><hr></div>
     <app-filter-preferences></app-filter-preferences>
