+++ frontend/src/app/features/settings/view-preferences-parent/view-preferences-parent.component.ts
@@ -6,8 +6,6 @@ import {ViewPreferencesComponent} from './view-preferences/view-preferences.comp
 import {SidebarSortingPreferencesComponent} from './sidebar-sorting-preferences/sidebar-sorting-preferences.component';
 import {MetaCenterViewModeComponent} from './meta-center-view-mode/meta-center-view-mode-component';
 import {FilterPreferencesComponent} from './filter-preferences/filter-preferences.component';
-import {LanguagePreferencesComponent} from './language-preferences/language-preferences.component';
-import {ThemePreferencesComponent} from './theme-preferences/theme-preferences.component';
 import {TranslocoDirective, TranslocoService} from '@jsverse/transloco';
 import {Slider} from 'primeng/slider';
 import {MessageService} from 'primeng/api';
@@ -23,8 +21,6 @@ import {LayoutService, SIDEBAR_MAX_WIDTH, SIDEBAR_MIN_WIDTH} from '../../../shar
     SidebarSortingPreferencesComponent,
     MetaCenterViewModeComponent,
     FilterPreferencesComponent,
-    LanguagePreferencesComponent,
-    ThemePreferencesComponent,
     TranslocoDirective,
     Slider,
   ],
