--- frontend/src/app/shared/layout/theme-configurator/theme-configurator.component.html
@@ -1,43 +1,0 @@
-<ng-container *transloco="let t; prefix: 'layout.theme'">
-  <div class="setting-item">
-    <div class="setting-header-standalone">
-      <span class="setting-label">{{ t('primary') }}</span>
-    </div>
-    <div class="setting-options">
-      <div class="color-swatches">
-        @for (primaryColor of primaryColors(); track primaryColor.name) {
-          <button
-            type="button"
-            [title]="primaryColor.name"
-            [attr.aria-label]="primaryColor.name"
-            [attr.aria-pressed]="primaryColor.name === selectedPrimaryColor()"
-            (click)="updateColors($event, 'primary', primaryColor)"
-            [ngClass]="{ 'active-color': primaryColor.name === selectedPrimaryColor() }"
-            [style]="{ 'background-color': primaryColor.name === 'noir' ? 'var(--text-color)' : primaryColor.palette['500'] || 'transparent' }"
-          ></button>
-        }
-      </div>
-    </div>
-  </div>
-
-  <div class="setting-item">
-    <div class="setting-header-standalone">
-      <span class="setting-label">{{ t('surface') }}</span>
-    </div>
-    <div class="setting-options">
-      <div class="color-swatches">
-        @for (surface of surfaces; track surface.name) {
-          <button
-            type="button"
-            [title]="surface.name"
-            [attr.aria-label]="surface.name"
-            [attr.aria-pressed]="selectedSurfaceColor() ? selectedSurfaceColor() === surface.name : surface.name === 'zinc'"
-            (click)="updateColors($event, 'surface', surface)"
-            [ngClass]="{ 'active-color': selectedSurfaceColor() ? selectedSurfaceColor() === surface.name : surface.name === 'zinc' }"
-            [style]="{'background-color': surface.name === 'noir' ? 'var(--text-color)' : surface.palette['500'] }">
-          </button>
-        }
-      </div>
-    </div>
-  </div>
-</ng-container>
