--- frontend/src/app/shared/layout/theme-configurator/theme-configurator.component.ts
@@ -1,61 +1,0 @@
-import {ChangeDetectionStrategy, Component, computed, inject} from '@angular/core';
-import {FormsModule} from '@angular/forms';
-import {AppConfigService} from '../../service/app-config.service';
-import {TranslocoDirective} from '@jsverse/transloco';
-import {NgClass} from '@angular/common';
-import Aura from '../theme-palette-extend';
-
-type ColorPalette = Record<string, string>;
-
-interface Palette {
-  name: string;
-  palette: ColorPalette;
-}
-
-@Component({
-  selector: 'app-theme-configurator',
-  standalone: true,
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './theme-configurator.component.html',
-  styleUrl: './theme-configurator.component.scss',
-  imports: [
-    NgClass,
-    FormsModule,
-    TranslocoDirective
-  ]
-})
-export class ThemeConfiguratorComponent {
-  readonly configService = inject(AppConfigService);
-
-  readonly surfaces = this.configService.surfaces;
-
-  readonly selectedPrimaryColor = computed(() => this.configService.appState().primary);
-  readonly selectedSurfaceColor = computed(() => this.configService.appState().surface);
-
-  readonly primaryColors = computed<Palette[]>(() => {
-    const presetPalette = (Aura.primitive ?? {}) as Record<string, ColorPalette>;
-    const colors = [
-      'orange', 'amber', 'yellow', 'lime', 'green', 'emerald', 
-      'teal', 'cyan', 'sky', 'blue', 'indigo', 'violet',
-      'purple', 'fuchsia', 'pink', 'rose', 'red',
-      'coralSunset', 'roseBlush', 'melonBlush', 'cottonCandy',
-      'apricotSunrise', 'antiqueBronze', 'butteryYellow', 'vanillaCream',
-      'citrusMint', 'freshMint', 'sagePearl', 'skyBlue','periwinkleCream',
-      'pastelRoyalBlue', 'lavenderDream', 'dustyNeutral'
-    ];
-    return [{name: 'noir', palette: {}}].concat(
-      colors.map(name => ({
-        name,
-        palette: presetPalette[name] ?? {}
-      }))
-    );
-  });
-
-  updateColors(event: Event, type: 'primary' | 'surface', color: { name: string; palette?: ColorPalette }) {
-    this.configService.appState.update((state) => ({
-      ...state,
-      [type]: color.name
-    }));
-    event.stopPropagation();
-  }
-}
